Refactor everything
Refactored the entire project.

This is more or less the new Initial Commit.
