+++ src/archive/chd.rs
@@ -1,0 +1,28 @@
+use std::error::Error;
+use std::path::Path;
+
+use crate::error::RomAnalyzerError;
+
+pub fn analyze_chd_file(_filepath: &Path, source_name: &str) -> Result<(), Box<dyn Error>> {
+    println!("\n=======================================================");
+    println!("  CHD ANALYSIS: Requires External Library (libchd)");
+    println!("=======================================================");
+
+    println!("In a real Rust environment, this function would use FFI (Foreign Function Interface) to bind to the MAME 'libchd' C library.");
+    println!("This library would decompress the hunks of data and extract the raw contents (e.g., a .BIN file) from the archive.");
+
+    // --- Conceptual Logic ---
+    // 1. FFI Call: chd_api::open(filepath) -> chd_handle
+    // 2. FFI Call: chd_api::read_raw_track(chd_handle) -> raw_data (Vec<u8>)
+
+    // As a placeholder, we will simulate the failure expected due to the lack of FFI,
+    // but clearly show the intended flow would route to the disc analysis.
+
+    // 3. Routing: process_rom_data(raw_data, virtual_filename)
+
+    // For demonstration, let's assume the CHD file contained PSX data.
+    // We cannot proceed without the external dependency.
+    return Err(Box::new(RomAnalyzerError::new(
+        &format!("CHD analysis for {} failed: FFI library 'libchd' is missing.", source_name)
+    )));
+}
